lots more frontend changes & cleaning up repo
