--- templates/css/style.css
@@ -1,7 +1,0 @@
-h1 {
-    color: blue;
-}
-
-.error-message {
-    color: red
-}
