+++ templates/index.html
@@ -1,9 +1,11 @@
 {% extends "base.html" %}
 {% block content %}
-<h1>This is the index page</h1> 
+<h1>Welcome to Citizen Cypher!</h1> 
 {% if session['loggedIn'] == True %}
 <h2>Welcome, {{ session['realname'] }}</h2>
+<p><a href="{{ url_for('logout')}}">Not you?</a></p>
 {% else %}
 <h2>You are not logged in.</h2>
+<p><a href="{{ url_for('login')}}">Log in</a> to your account or <a href="{{ url_for('signup')}}">sign up</a> for an account!</p>
 {% endif %}
 {% endblock %}
