--- templates/js/index.js
@@ -1,2 +1,0 @@
-console.log('Hello World!');
-
