+++ templates/login.html
@@ -15,7 +15,7 @@
                     <img src="/static/img/avatar.png">
                 </div>
                 <br>
-                <div class="welcomeMessage">
+                <div class="welcomeMessage box-text">
                     <h1>Login to Citizen Cypher</h1>
                     <h3>Please enter your credentials.</h3>
                 </div>
@@ -34,7 +34,7 @@
                 </form>
             </div>
             <div class="lowerBox">
-                <h4>If you don't have an account, sign up here:</h4>
+                <h4 class="box-text">If you don't have an account, sign up here:</h4>
                 <div class="signup">
                     <!-- <button type="submit" class="sign-btn button">SIGN UP</button> -->
                     <button onclick="window.location.href='./signup';" type="submit" class="sign-btn button">SIGN UP</button>
