+++ templates/signup.html
@@ -16,7 +16,7 @@
                     <img src="/static/img/avatar.png">
                 </div>
                 <br>
-                <div class="welcomeMessage">
+                <div class="welcomeMessage box-text">
                     <h1>Sign up for Citizen Cypher</h1>
                     <h3>To create an account, enter your credentials!</h3>
                 </div>
@@ -40,7 +40,7 @@
                 
             </div>
             <div class="lowerBox">
-                <h4>If you already have an account, login instead:</h4>
+                <h4 class="box-text">If you already have an account, login instead:</h4>
                 <div class="login">
                     <!-- <button type="submit" class="sign-btn button">SIGN UP</button> -->
                     <button onclick="window.location.href='./login';" type="submit" class="login-btn button">LOGIN</button>
